Moved the startTimePicker into its own component
